+++ working-php/src/Views/partials/signing-modal.php
@@ -1,6 +1,5 @@
 <!-- Digital Signature Modal -->
 <?php
-$hasCachedPin = \App\Core\SecurityHelper::hasCachedPin();
 
 $modalId = 'signing-modal';
 $modalTitle = 'Digital Signature';